% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{RCSmixed.fit}
\alias{RCSmixed.fit}
\title{A function for implementing the REMARCS algorithm with mixed effect model}
\usage{
RCSmixed.fit(mod, dat, resp_var, tol = 0.1)
}
\arguments{
\item{mod}{The initial model (mixed lmer) of the response variable}

\item{dat}{The data in question. The variable for time must be entitled "t".}

\item{resp_var}{The name of the variable being modeled}

\item{tol}{Convergence tolerance for the iterative estimation procedure.}
}
\description{
A function for implementing the REMARCS algorithm with mixed effect model
}
