% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{RCSreg.fit}
\alias{RCSreg.fit}
\title{Fit the REMARCS model with a fixed-effects regression component}
\usage{
RCSreg.fit(mod, dat, resp_var, tol = 0.1)
}
\arguments{
\item{mod}{The initial fixed-effects model of the response variable.}

\item{dat}{The data used for fitting. The variable for time must be named
\code{t}.}

\item{resp_var}{Character string giving the name of the response variable.}

\item{tol}{Convergence tolerance for the iterative REMARCS algorithm.}
}
\value{
A list containing the estimated regression parameters, time-series
parameters, log-likelihood, AIC, and fitted values.
}
\description{
Implements the REMARCS algorithm using an initial fixed-effects model
and a latent ARMA(1, 1) time-series component.
}
