% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\docType{package}
\name{remarcs-package}
\alias{remarcs}
\alias{remarcs-package}
\title{remarcs: REMARCS Estimation for Repeated Cross-Sectional Data}
\description{
Tools for estimating repeated cross-sectional regression models with
latent temporal dependence using the REMARCS algorithm.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
