% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{RCSmixed.fit}
\alias{RCSmixed.fit}
\title{Fit the REMARCS model with a mixed-effects regression component}
\usage{
RCSmixed.fit(mod, dat, resp_var, tol = 0.1, max_updates = 10)
}
\arguments{
\item{mod}{The initial mixed-effects model fitted with \code{lme4::lmer()}.}

\item{dat}{The data used for fitting. The variable for time must be named
\code{t}.}

\item{resp_var}{Character string giving the name of the response variable.}

\item{tol}{Convergence tolerance for the iterative REMARCS algorithm.}
}
\value{
A list containing the estimated regression parameters, time-series
parameters, log-likelihood, AIC, and fitted values.
}
\description{
Implements the REMARCS algorithm using an initial mixed-effects model
and a latent ARMA(1, 1) time-series component.
}
